refactor: migrate household feature to feature slice
- Create src/features/household/ with:
  - context.ts - HouseholdContext and HouseholdContextValue type
  - provider.tsx - HouseholdProvider component
  - hooks/useHousehold.ts - useHousehold hook
  - utils/calculations.ts - calculateHouseholdMultiplier, calculateRecommendedQuantity
  - utils/calculations.test.ts - unit tests for calculations
  - constants.ts - HOUSEHOLD_DEFAULTS, HOUSEHOLD_LIMITS
  - index.ts - public API (barrel exports)

- Update all imports across codebase to use @/features/household

- Create backward-compatible re-exports in shared code:
  - src/shared/hooks/useHousehold.ts
  - src/shared/contexts/HouseholdContext.ts
  - src/shared/contexts/HouseholdProvider.tsx
  - src/shared/constants/household.ts
  - src/shared/utils/calculations/household.ts

All 888 tests pass.
