+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { trackAppLaunch } from '@/shared/utils/analytics';
 import { SettingsProvider } from '@/shared/contexts/SettingsProvider';
-import { HouseholdProvider } from '@/shared/contexts/HouseholdProvider';
+import { HouseholdProvider, useHousehold } from '@/features/household';
 import { InventoryProvider } from '@/shared/contexts/InventoryProvider';
 import { RecommendedItemsProvider } from '@/shared/contexts/RecommendedItemsProvider';
 import { ThemeApplier } from './components/ThemeApplier';
@@ -14,7 +14,6 @@ import { Settings } from './pages/Settings';
 import { Help } from './pages/Help';
 import { Onboarding } from './components/onboarding/Onboarding';
 import { useSettings } from '@/shared/hooks/useSettings';
-import { useHousehold } from '@/shared/hooks/useHousehold';
 import { useInventory } from '@/shared/hooks/useInventory';
 import type { HouseholdConfig, InventoryItem } from '@/shared/types';
 import './App.css';
