+++ src/components/settings/HiddenAlerts.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useInventory } from '@/shared/hooks/useInventory';
-import { useHousehold } from '@/shared/hooks/useHousehold';
+import { useHousehold } from '@/features/household';
 import { generateDashboardAlerts } from '@/shared/utils/dashboard/alerts';
 import { Button } from '@/shared/components/Button';
 import styles from './HiddenAlerts.module.css';
