+++ src/features/household/context.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+import type { HouseholdConfig } from '@/shared/types';
+
+export interface HouseholdContextValue {
+  household: HouseholdConfig;
+  updateHousehold: (updates: Partial<HouseholdConfig>) => void;
+  setPreset: (preset: 'single' | 'couple' | 'family') => void;
+}
+
+export const HouseholdContext = createContext<
+  HouseholdContextValue | undefined
+>(undefined);
