+++ src/features/household/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useHousehold } from './useHousehold';
