+++ src/features/household/hooks/useHousehold.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { HouseholdContext } from '../context';
+
+export function useHousehold() {
+  const context = useContext(HouseholdContext);
+  if (!context) {
+    throw new Error('useHousehold must be used within HouseholdProvider');
+  }
+  return context;
+}
