+++ src/features/household/index.ts
@@ -1,0 +1,16 @@
+// Context and Provider
+export { HouseholdContext } from './context';
+export type { HouseholdContextValue } from './context';
+export { HouseholdProvider } from './provider';
+
+// Hooks
+export { useHousehold } from './hooks';
+
+// Utils
+export {
+  calculateHouseholdMultiplier,
+  calculateRecommendedQuantity,
+} from './utils';
+
+// Constants
+export { HOUSEHOLD_DEFAULTS, HOUSEHOLD_LIMITS } from './constants';
