+++ src/features/household/utils/index.ts
@@ -1,0 +1,4 @@
+export {
+  calculateHouseholdMultiplier,
+  calculateRecommendedQuantity,
+} from './calculations';
