+++ src/main.tsx
@@ -3,8 +3,8 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import './i18n/config';
 import App from './App.tsx';
+import { HouseholdProvider } from '@/features/household';
 import { InventoryProvider } from '@/shared/contexts/InventoryProvider';
-import { HouseholdProvider } from '@/shared/contexts/HouseholdProvider';
 import { SettingsProvider } from '@/shared/contexts/SettingsProvider';
 import { ErrorBoundary } from '@/shared/components/ErrorBoundary';
 import { ThemeApplier } from './components/ThemeApplier';
