+++ src/pages/Dashboard.tsx
@@ -6,7 +6,7 @@ import type { Alert } from '@/shared/utils/dashboard/alerts';
 import { CategoryGrid } from '../components/dashboard/CategoryGrid';
 import { Button } from '@/shared/components/Button';
 import { useInventory } from '@/shared/hooks/useInventory';
-import { useHousehold } from '@/shared/hooks/useHousehold';
+import { useHousehold } from '@/features/household';
 import { useSettings } from '@/shared/hooks/useSettings';
 import { useRecommendedItems } from '@/shared/hooks/useRecommendedItems';
 import { STANDARD_CATEGORIES } from '@/data/standardCategories';
