+++ src/shared/hooks/useHousehold.ts
@@ -1,10 +1,2 @@
-import { useContext } from 'react';
-import { HouseholdContext } from '@/shared/contexts/HouseholdContext';
-
-export function useHousehold() {
-  const context = useContext(HouseholdContext);
-  if (!context) {
-    throw new Error('useHousehold must be used within HouseholdProvider');
-  }
-  return context;
-}
+// Re-export from feature slice for backward compatibility
+export { useHousehold } from '@/features/household';
